fix unsigned int and signed int compare warning and pow under vc9 no int64 version bug.

+++ cpp/adata_int64.hpp
@@ -434,7 +434,7 @@ namespace adata {
         }
         case et_int64_uint64:
         {
-          lua_pushboolean(L, it1.value.i64 < it2.value.u64);
+          lua_pushboolean(L, it1.value.i64 < (double)it2.value.u64);
           break;
         }
         default:
@@ -455,7 +455,7 @@ namespace adata {
         }
         case et_int64_int64:
         {
-          lua_pushboolean(L, it1.value.u64 < it2.value.i64);
+          lua_pushboolean(L, (double)it1.value.u64 < it2.value.i64);
           break;
         }
         case et_int64_uint64:
@@ -527,7 +527,7 @@ namespace adata {
         }
         case et_int64_uint64:
         {
-          lua_pushboolean(L, it1.value.i64 <= it2.value.u64);
+          lua_pushboolean(L, it1.value.i64 <= (double)it2.value.u64);
           break;
         }
         default:
@@ -548,7 +548,7 @@ namespace adata {
         }
         case et_int64_int64:
         {
-          lua_pushboolean(L, it1.value.u64 <= it2.value.i64);
+          lua_pushboolean(L, (double)it1.value.u64 <= it2.value.i64);
           break;
         }
         case et_int64_uint64:
@@ -1054,12 +1054,12 @@ namespace adata {
         }
         case et_int64_int64:
         {
-          lua_pushnumber(L, pow(it1.value.d64, it2.value.i64));
+          lua_pushnumber(L, pow(it1.value.d64, (double)it2.value.i64));
           break;
         }
         case et_int64_uint64:
         {
-          lua_pushnumber(L, pow(it1.value.d64, it2.value.u64));
+          lua_pushnumber(L, pow(it1.value.d64, (double)it2.value.u64));
           break;
         }
         default:
@@ -1080,7 +1080,7 @@ namespace adata {
         }
         case et_int64_int64:
         {
-          lua_pushint64(L, (int64_t)pow(it1.value.i64, it2.value.i64));
+          lua_pushint64(L, (int64_t)pow((double)it1.value.i64, (double)it2.value.i64));
           break;
         }
         case et_int64_uint64:
@@ -1106,12 +1106,12 @@ namespace adata {
         }
         case et_int64_int64:
         {
-          lua_pushuint64(L, (uint64_t)pow(it1.value.u64, it2.value.i64));
+          lua_pushuint64(L, (uint64_t)pow((double)it1.value.u64, (double)it2.value.i64));
           break;
         }
         case et_int64_uint64:
         {
-          lua_pushuint64(L, (uint64_t)pow(it1.value.u64, it2.value.u64));
+          lua_pushuint64(L, (uint64_t)pow((double)it1.value.u64, (double)it2.value.u64));
           break;
         }
         default:
